Add files via upload
Added functionality for Transfer and menu GUIs
The transfer doesn't seem to change the account balances and I don't have a way to check the balance before transferring
The MenuGUI has issues with the history still.  It looks like the interest calculator is working- had to bypass the Main.java and call the interest calculator directly and it was a pain converting from string to double and double to string.